rename litecoin-api to litecoin
