+++ api/litecoin.js
